requirements.txt
absl-py==0.10.0
adal==1.2.4
aiohttp==3.6.2
applicationinsights==0.11.9
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
azure-common==1.1.25
azure-core==1.8.1
azure-graphrbac==0.61.1
azure-identity==1.4.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.15.0
azureml-core==1.15.0
azureml-dataprep==2.2.3
azureml-dataprep-native==22.0.0
azureml-dataprep-rslex==1.0.1
azureml-dataset-runtime==1.15.0
azureml-pipeline==1.15.0
azureml-pipeline-core==1.15.0
azureml-pipeline-steps==1.15.0
azureml-sdk==1.15.0
azureml-telemetry==1.15.0
azureml-train==1.15.0
azureml-train-automl-client==1.15.0.post1
azureml-train-core==1.15.0
azureml-train-restclients-hyperdrive==1.15.0
backports.tempfile==1.0
backports.weakref==1.0.post1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.1.1
distro==1.5.0
docker==4.3.1
dotnetcore2==2.1.16
fusepy==3.0.1
fuzzywuzzy==0.18.0
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
isodate==0.6.0
jeepney==0.4.3
jmespath==0.10.0
joblib==0.17.0
jsonpickle==1.4.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
msal==1.5.0
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
multidict==4.7.6
ndg-httpsclient==0.5.1
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pathspec==0.8.0
portalocker==1.7.1
protobuf==3.13.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.2
SecretStorage==3.1.2
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
textblob==0.15.3
tqdm==4.50.0
typing-extensions==3.7.4.3
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.3.0
